Make star ratings an independent function
Previously, the DOM elements that display the rating using stars were
created in the displayRating function, which also had the logic for
displaying the numerical rating.

Drawing stars has now been moved into makeStarRating, which now also
returns the created star container, instead of adding it via function
side effects. This is because the stars will need to be reused when
displaying ratings in user reviews.